Adopts non-destructive prebuild script.
We were removing the build directory entirely before making
it. Now, we’re only making the directory if needed. It's a
much more polite approach, though it could bite if we don’t
test things in a container first. I can imagine a scenario
where we think we’re rebuilding a file, but it’s really
just an old version of that file that’s been sitting in our
local, and we don’t notice until we’re missing the file
downstream.